# src/main/resources/static/swagger-api/swagger.yaml
openapi: 3.0.0
info:
  version: "1.0.0"
  title: "CUMyBank"
servers:
  - url: "http://localhost:8080"
...
paths:
...
  /api/v1/admin/user/all:
    get:
      tags:
        - "AdminController"
      summary: "Get users from database"
      operationId: "getAllUsers"
      parameters: []
      responses:
        "200":
          description: "OK"
          content:
            'application/json':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
      security:
        - token: []
...
  /api/v1/registration/activate/{code}:
    get:
      tags:
        - "RegistrationController"
      summary: "Activate email code"
      operationId: "activateEmailCode"
      parameters:
        - name: "code"
          in: "path"
          required: true
          schema:
            type: "string"
      responses:
        "200":
          description: "OK"
          content:
            'application/json':
              schema:
                type: object
                properties:
                  data:
                    type: "string"
                    example: "User successfully activated."
        "404":
          description: "Bad Request"
          content:
            'application/json':
              schema:
                type: object
                properties:
                  data:
                    type: string
                    example: "Activation code not found."
...
  /api/v1/account/new/:
    post:
      tags:
        - "AccountController"
      summary: "Create account by given parameters"
      operationId: "createAccountByGivenParameters"
      requestBody:
        required: true
        content:
          'application/json':
            schema:
              type: object
              properties:
                number:
                  type: "string"
                  example: "138738495720478"
                currency:
                  type: "string"
                  example: "RUB"
                type:
                  type: "string"
                  example: "CREDIT"
                paymentSystem:
                  type: "string"
                  example: "VISA"
                mail:
                  type: "string"
                  example: "example@gmail.com"
      responses:
        "200":
          description: "OK"
          content:
            'application/json':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TransferResponse"
components:
  securitySchemes:
    token:
      type: apiKey
      name: Authorization
      in: header
  schemas:
    TransferResponse:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int64"
          example: 43
        dateTime:
          type: "string"
          example: "2021-04-07"
        sender:
          type: "string"
          example: "Dior"
        recipient:
          type: "string"
          example: "Dior"
        status:
          type: "string"
          example: "France"
        commission:
          type: "number"
          format: "double"
          example: "0.1"
        amount:
          type: "number"
          format: "double"
          example: "100.10"
...
    User:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int64"
          example: 2
        email:
          type: "string"
          example: "test123@test.com"
        firstName:
          type: "string"
          example: "John"
        lastName:
          type: "string"
          example: "Doe"
        active:
          type: "boolean"
          example: true
        activationCode:
          type: "string"
          example: ""
        passwordResetCode:
          type: "string"
          example: ""
        role:
          type: "string"
          example: "USER"
...